# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/moneytransfers": {
            "get": {
                "description": "list moneytransfer entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List moneytransfer entities",
                "operationId": "list-moneytransfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.MoneyTransfer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create moneytransfer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create moneytransfer",
                "operationId": "create-moneytransfer",
                "parameters": [
                    {
                        "description": "MoneyTransfer entity",
                        "name": "moneyTransfer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.MoneyTransfer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.MoneyTransfer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/moneytransfers/{id}": {
            "get": {
                "description": "get moneytransfer by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a moneytransfer entity by ID",
                "operationId": "get-moneytransfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MoneyTransfer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.MoneyTransfer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update moneytransfer by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a moneytransfer entity by ID",
                "operationId": "update-moneytransfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MoneyTransfer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MoneyTransfer entity",
                        "name": "moneyTransfer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.MoneyTransfer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.MoneyTransfer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get moneytransfer by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a moneytransfer entity by ID",
                "operationId": "delete-moneytransfer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "MoneyTransfer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Bank": {
            "type": "object",
            "properties": {
                "bank_type": {
                    "description": "BankType holds the value of the \"bank_type\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BankQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BankEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BankEdges": {
            "type": "object",
            "properties": {
                "bankPayback": {
                    "description": "BankPayback holds the value of the bank_payback edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payback"
                    }
                },
                "bankPayment": {
                    "description": "BankPayment holds the value of the bank_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                }
            }
        },
...
        "ent.Gender": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "gender_name": {
                    "description": "GenderName holds the value of the \"gender_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Hospital": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the HospitalQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.HospitalEdges"
                },
                "hospital_name": {
                    "description": "HospitalName holds the value of the \"hospital_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Inquiry": {
            "type": "object",
            "properties": {
                "Inquiry_messages": {
                    "description": "InquiryMessages holds the value of the \"Inquiry_messages\" field.",
                    "type": "string"
                },
                "Inquiry_time_messages": {
                    "description": "InquiryTimeMessages holds the value of the \"Inquiry_time_messages\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the InquiryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.InquiryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Insurance": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the InsuranceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.InsuranceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_address": {
                    "description": "InsuranceAddress holds the value of the \"insurance_address\" field.",
                    "type": "string"
                },
                "insurance_insurer": {
                    "description": "InsuranceInsurer holds the value of the \"insurance_insurer\" field.",
                    "type": "string"
                },
                "insurance_time_buy": {
                    "description": "InsuranceTimeBuy holds the value of the \"insurance_time_buy\" field.",
                    "type": "string"
                },
                "insurance_time_firstpay": {
                    "description": "InsuranceTimeFirstpay holds the value of the \"insurance_time_firstpay\" field.",
                    "type": "string"
                }
            }
        },
        "ent.InsuranceEdges": {
            "type": "object",
            "properties": {
                "hospital": {
                    "description": "Hospital holds the value of the Hospital edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Hospital"
                },
                "insurancePayment": {
                    "description": "InsurancePayment holds the value of the insurance_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                },
                "member": {
                    "description": "Member holds the value of the Member edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Member"
                },
                "officer": {
                    "description": "Officer holds the value of the Officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "product": {
                    "description": "Product holds the value of the Product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                }
            }
        },
        "ent.Member": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MemberQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MemberEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "member_email": {
                    "description": "MemberEmail holds the value of the \"member_email\" field.",
                    "type": "string"
                },
                "member_name": {
                    "description": "MemberName holds the value of the \"member_name\" field.",
                    "type": "string"
                },
                "member_password": {
                    "description": "MemberPassword holds the value of the \"member_password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MemberEdges": {
            "type": "object",
            "properties": {
                "memberInquiry": {
                    "description": "MemberInquiry holds the value of the member_inquiry edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Inquiry"
                    }
                },
                "memberInsurance": {
                    "description": "MemberInsurance holds the value of the member_insurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Insurance"
                    }
                },
                "memberPayback": {
                    "description": "MemberPayback holds the value of the member_payback edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payback"
                    }
                },
                "memberPayment": {
                    "description": "MemberPayment holds the value of the member_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                },
                "memberRecordinsurance": {
                    "description": "MemberRecordinsurance holds the value of the member_recordinsurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Recordinsurance"
                    }
                }
            }
        },
        "ent.MoneyTransfer": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MoneyTransferQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MoneyTransferEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "moneytransfer_type": {
                    "description": "MoneytransferType holds the value of the \"moneytransfer_type\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MoneyTransferEdges": {
            "type": "object",
            "properties": {
                "moneytransferPayment": {
                    "description": "MoneytransferPayment holds the value of the moneytransfer_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                }
            }
        },
        "ent.Officer": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OfficerQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OfficerEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "officer_email": {
                    "description": "OfficerEmail holds the value of the \"officer_email\" field.",
                    "type": "string"
                },
                "officer_name": {
                    "description": "OfficerName holds the value of the \"officer_name\" field.",
                    "type": "string"
                },
                "officer_password": {
                    "description": "OfficerPassword holds the value of the \"officer_password\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Payback": {
            "type": "object",
            "properties": {
                "Accountnumber": {
                    "description": "Accountnumber holds the value of the \"Accountnumber\" field.",
                    "type": "string"
                },
                "Transfertime": {
                    "description": "Transfertime holds the value of the \"Transfertime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaybackQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaybackEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Payment": {
            "type": "object",
            "properties": {
                "account_name": {
                    "description": "AccountName holds the value of the \"account_name\" field.",
                    "type": "string"
                },
                "account_number": {
                    "description": "AccountNumber holds the value of the \"account_number\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaymentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "transfer_time": {
                    "description": "TransferTime holds the value of the \"transfer_time\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PaymentEdges": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "Bank holds the value of the Bank edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bank"
                },
                "insurance": {
                    "description": "Insurance holds the value of the Insurance edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Insurance"
                },
                "member": {
                    "description": "Member holds the value of the Member edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Member"
                },
                "moneyTransfer": {
                    "description": "MoneyTransfer holds the value of the MoneyTransfer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MoneyTransfer"
                },
                "product": {
                    "description": "Product holds the value of the Product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                }
            }
        },
        "ent.Product": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProductQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProductEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "product_name": {
                    "description": "ProductName holds the value of the \"product_name\" field.",
                    "type": "string"
                },
                "product_payment_of_year": {
                    "description": "ProductPaymentOfYear holds the value of the \"product_payment_of_year\" field.",
                    "type": "number"
                },
                "product_price": {
                    "description": "ProductPrice holds the value of the \"product_price\" field.",
                    "type": "integer"
                },
                "product_time": {
                    "description": "ProductTime holds the value of the \"product_time\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.ProductEdges": {
            "type": "object",
            "properties": {
                "productGender": {
                    "description": "ProductGender holds the value of the product_gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "productGroupage": {
                    "description": "ProductGroupage holds the value of the product_groupage edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GroupOfAge"
                },
                "productInquiry": {
                    "description": "ProductInquiry holds the value of the product_inquiry edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Inquiry"
                    }
                },
                "productInsurance": {
                    "description": "ProductInsurance holds the value of the product_insurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Insurance"
                    }
                },
                "productOfficer": {
                    "description": "ProductOfficer holds the value of the product_officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "productPayback": {
                    "description": "ProductPayback holds the value of the product_payback edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payback"
                    }
                },
                "productPayment": {
                    "description": "ProductPayment holds the value of the product_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                },
                "productRecordinsurance": {
                    "description": "ProductRecordinsurance holds the value of the product_recordinsurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Recordinsurance"
                    }
                }
            }
        },
        "ent.Recordinsurance": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RecordinsuranceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RecordinsuranceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "recordinsurance_time": {
                    "description": "RecordinsuranceTime holds the value of the \"recordinsurance_time\" field.",
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}